+++ src/lib/terkin/device.py
@@ -40,11 +40,11 @@ class TerkinDevice:
         self.watchdog = Watchdog(device=self, settings=self.settings)
 
         # Conditionally enable terminal on UART0. Default: False.
-        try:
-            self.terminal = Terminal(self.settings)
-            self.terminal.start()
-        except Exception as ex:
-            log.exc(ex, 'Enabling Terminal failed')
+        #try:
+        #    self.terminal = Terminal(self.settings)
+        #    self.terminal.start()
+        #except Exception as ex:
+        #    log.exc(ex, 'Enabling Terminal failed')
 
         self.device_id = get_device_id()
 
@@ -103,6 +103,17 @@ class TerkinDevice:
                 log.info("[LoRa] interface disabled in settings. Skip LoRa initialisation.")
         else:
             log.info("[LoRa] This is not a LoRa capable device.")
+
+        # Initialize GPRS modem.
+        if self.settings.get('networking.gprs.enabled'):
+            try:
+                self.networking.start_gprs()
+                self.status.networking = True
+            except Exception as ex:
+                log.exc(ex, 'Unable to start GPRS modem')
+                self.status.networking = False
+        else:
+            log.info("[GPRS] Interface disabled in settings, skipping initialization.")
 
         # Inform about networking status.
         #self.networking.print_status()
@@ -372,7 +383,7 @@ class TerkinDevice:
 
             # Invoke deep sleep.
             log.info('Entering deep sleep for {} seconds'.format(interval))
-            self.terminal.stop()
+            #self.terminal.stop()
             machine.deepsleep(int(interval * 1000))
 
         else: